+++ Project1/index.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <string>
 int main(int argc, char *argv[]);
-bool isSlash(std::string filename, int position);
+bool isSlash(std::string filename, size_t position);
